+++ src/Utils/Utils.php
@@ -6,6 +6,9 @@ use function array_keys;
 use function array_map;
 use function array_reduce;
 use function array_slice;
+
+use ArrayAccess;
+
 use function count;
 use function dechex;
 use function get_class;
@@ -314,5 +317,25 @@ class Utils
         });
 
         return array_map('strval', array_keys($optionsByDistance));
+    }
+
+    /**
+     * Try to extract the value for a key from an object like value.
+     *
+     * @param mixed $objectLikeValue
+     *
+     * @return mixed
+     */
+    public static function extractKey($objectLikeValue, string $key)
+    {
+        if (is_array($objectLikeValue) || $objectLikeValue instanceof ArrayAccess) {
+            return $objectLikeValue[$key] ?? null;
+        }
+
+        if (is_object($objectLikeValue)) {
+            return $objectLikeValue->{$key} ?? null;
+        }
+
+        return null;
     }
 }