+++ CHANGELOG.md
@@ -71,6 +71,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Allow constructing `EnumType` from PHP enum
 - Add `TypeInfo::getParentTypeStack()` and `TypeInfo::getFieldDefStack()`
 - Include path to faulty input in coercion errors
+- Add ability to resolve abstract type of object via `__typename`
 
 ### Optimized